long policy
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@5929 a1c6a512-1295-4272-9138-f99709370657

+++ apps/settings.h
@@ -296,7 +296,7 @@ enum optiontype { INT, BOOL };
 
 struct opt_items {
     unsigned const char* string;
-    int voice_id;
+    long voice_id;
 };
 
 /* prototypes */
